# Conflicts:
#	azure-pipelines.yml
#	src/MockResponses/setup.json
#	src/NBS.Appointments.Service.Api.Tests/HelloWorldApiTests.cs
#	src/NBS.Appointments.Service.Core/Dtos/Qflow/AvailabilityResponse.cs
#	src/NBS.Appointments.Service.Core/Dtos/Qflow/QflowServiceDescriptor.cs
#	src/NBS.Appointments.Service.Core/Interfaces/Services/IQflowService.cs
#	src/NBS.Appointments.Service.Core/NBS.Appointments.Service.Core.csproj
#	src/NBS.Appointments.Service.Core/ServiceRegistration.cs
#	src/NBS.Appointments.Service.Core/Services/QflowService.cs
#	src/NBS.Appointments.Service.sln
#	src/NBS.Appointments.Service/Controllers/AvailabilityController.cs
#	src/NBS.Appointments.Service/Models/Availability.cs
#	src/NBS.Appointments.Service/NBS.Appointments.Service.csproj
#	src/NBS.Appointments.Service/Program.cs
#	src/NBS.Appointments.Service/Properties/launchSettings.json
#	src/NBS.Appointments.Service/Startup.cs
#	src/NBS.Appointments.Service/appsettings.Development.json
#	src/docker-compose.yml
#	terraform/resources/api.tf
#	terraform/resources/locals.tf
#	terraform/resources/variables.tf

+++ src/NBS.Appointments.Service.Core/Dtos/Qflow/AvailabilityResponse.cs
@@ -7,9 +7,9 @@ namespace NBS.Appointments.Service.Infrastructure.Dtos.Qflow
         public DateTime Date { get; set; }
 
         [JsonProperty("AM")]
-        public long Am { get; set; }
+        public int Am { get; set; }
 
         [JsonProperty("PM")]
-        public long Pm { get; set; }
+        public int Pm { get; set; }
     }
 }
